+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/listener/TaskLevelListener.java
@@ -8,10 +8,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import lombok.Getter;
 import org.apache.spark.SparkContext;
-import org.apache.spark.executor.TaskMetrics;
 import org.apache.spark.scheduler.SparkListenerTaskEnd;
-import org.apache.spark.scheduler.SparkListenerTaskStart;
-import org.apache.spark.scheduler.TaskInfo;
 
 /**
  * This class is a task-level listener for the Spark data source.
@@ -40,21 +37,6 @@ public class TaskLevelListener extends TaskStageBaseListener {
     super(sparkContext, config);
   }
 
-  @Override
-  public void onTaskStart(SparkListenerTaskStart taskStart) {
-    final long taskId = taskStart.taskInfo().taskId() + 1;
-    final int stageId = taskStart.stageId();
-    final List<Long> tasks = stageToTasks.get(stageId);
-    if (tasks == null) {
-      List<Long> newTasks = new ArrayList<>();
-      newTasks.add(taskId);
-      stageToTasks.put(stageId, newTasks);
-    } else {
-      tasks.add(taskId);
-    }
-    taskToStage.put(taskId, stageId);
-  }
-
   /**
    * This method is called every time a task ends, task-level metrics should be collected here, and added.
    *
@@ -65,19 +47,26 @@ public class TaskLevelListener extends TaskStageBaseListener {
    */
   @Override
   public void onTaskEnd(SparkListenerTaskEnd taskEnd) {
-    final TaskInfo curTaskInfo = taskEnd.taskInfo();
-    final TaskMetrics curTaskMetrics = taskEnd.taskMetrics();
-    final long taskId = curTaskInfo.taskId() + 1;
+    final long taskId = taskEnd.taskInfo().taskId() + 1;
+    final int stageId = taskEnd.stageId() + 1;
+    taskToStage.put(taskId, stageId);
+
+    final List<Long> tasks = stageToTasks.get(stageId);
+    if (tasks == null) {
+      List<Long> newTasks = new ArrayList<>();
+      newTasks.add(taskId);
+      stageToTasks.put(stageId, newTasks);
+    } else {
+      tasks.add(taskId);
+    }
 
     final String type = taskEnd.taskType();
-    final long submitTime = curTaskInfo.launchTime();
-    final long runTime = curTaskMetrics.executorRunTime();
-    final int userId = sparkContext.sparkUser().hashCode();
-
-    final long workflowId = stageIdsToJobs.get(taskEnd.stageId() + 1);
-
+    final long tsSubmit = taskEnd.taskInfo().launchTime();
+    final long runtime = taskEnd.taskMetrics().executorRunTime();
     final long[] parents = new long[0];
     final long[] children = new long[0];
+    final int userId = sparkContext.sparkUser().hashCode();
+    final long workflowId = stageIdsToJobs.get(stageId);
 
     // unknown
     final int submissionSite = -1;
@@ -101,8 +90,8 @@ public class TaskLevelListener extends TaskStageBaseListener {
             .id(taskId)
             .type(type)
             .submissionSite(submissionSite)
-            .tsSubmit(submitTime)
-            .runtime(runTime)
+            .tsSubmit(tsSubmit)
+            .runtime(runtime)
             .resourceType(resourceType)
             .resourceAmountRequested(resourceAmountRequested)
             .parents(parents)