+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/EndToEnd.java
@@ -1,9 +1,12 @@
 package com.asml.apa.wta.spark;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.spark.SparkConf;
 import org.apache.spark.SparkContext;
+import org.apache.spark.api.java.JavaPairRDD;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.sql.SparkSession;
@@ -19,19 +22,112 @@ import scala.Tuple2;
 @Slf4j
 public class EndToEnd {
 
-  private static JavaRDD<String> testFile;
-
   /**
-   * Private method to invoke a simple Spark job.
-   *
+   * Private method to invoke the Spark application with complex jobs involving partitions shuffling. Results don't
+   * matter as the purpose is to generate Spark tasks with multiple parent-child relations and jobs with diverse
+   * number of tasks for generated traces.
+   * @param textFile  The text file to read from.
    * @author Pil Kyu Cho
    * @since 1.0.0
    */
-  private static void invokeJob() {
-    testFile.flatMap(s -> Arrays.asList(s.split(" ")).iterator())
-        .mapToPair(word -> new Tuple2<>(word, 1))
+  private static void sparkOperation(JavaRDD<String> textFile) {
+    JavaRDD<String> words1 = textFile.flatMap(
+            line -> Arrays.asList(line.split(" ")).iterator())
+        .filter(word -> word.contains("harry"))
+        .map(String::toUpperCase);
+    JavaRDD<String> words2 = textFile.flatMap(
+            line -> Arrays.asList(line.split(" ")).iterator())
+        .filter(word -> word.contains("potter"))
+        .map(String::toUpperCase);
+    JavaPairRDD<String, Tuple2<Integer, Integer>> wordCountPairs1 = words1.mapToPair(word -> new Tuple2<>(word, 1))
         .reduceByKey(Integer::sum)
-        .collect();
+        .mapToPair(tuple ->
+            new Tuple2<>(tuple._1(), new Tuple2<>(tuple._1().length(), tuple._2())));
+    JavaPairRDD<String, Tuple2<Integer, Integer>> wordCountPairs2 = words2.mapToPair(word -> new Tuple2<>(word, 1))
+        .reduceByKey(Integer::sum)
+        .mapToPair(tuple ->
+            new Tuple2<>(tuple._1(), new Tuple2<>(tuple._1().length(), tuple._2())));
+    JavaPairRDD<Tuple2<String, String>, Tuple2<Integer, Integer>> joinedPairs = wordCountPairs1
+        .cartesian(wordCountPairs2)
+        .filter(tuple -> tuple._1()._1().compareTo(tuple._2()._1()) < 0)
+        .mapToPair(tuple -> new Tuple2<>(
+            new Tuple2<>(tuple._1()._1(), tuple._2()._1()),
+            new Tuple2<>(
+                tuple._1()._2()._1() + tuple._2()._2()._1(),
+                tuple._1()._2()._2() + tuple._2()._2()._2())));
+    JavaPairRDD<Double, Tuple2<String, String>> sortedPairs = joinedPairs
+        .filter(tuple -> tuple._2()._2() > 3)
+        .mapValues(tuple -> (double) tuple._1() / tuple._2())
+        .mapToPair(pair -> new Tuple2<>(pair._2(), pair._1()))
+        .repartition(10)
+        .sortByKey(false);
+    JavaRDD<Tuple2<Double, Tuple2<String, String>>> tupled = sortedPairs.mapPartitionsWithIndex(
+        (index, iter) -> {
+          List<Tuple2<Double, Tuple2<String, String>>> tuples = new ArrayList<>();
+          while (iter.hasNext()) {
+            tuples.add(new Tuple2<>(
+                iter.next()._1(),
+                new Tuple2<>(
+                    index + "-" + iter.next()._2()._1(),
+                    iter.next()._2()._2())));
+          }
+          return tuples.iterator();
+        },
+        true);
+
+    JavaRDD<Tuple2<Double, String>> mappedPairs = tupled.mapToPair(tuple -> new Tuple2<>(tuple._1(), tuple._2()))
+        .groupByKey()
+        .filter(pair -> pair._1() >= 10)
+        .map(pair -> {
+          StringBuilder concatenated = new StringBuilder();
+          for (Tuple2<String, String> value : pair._2()) {
+            concatenated
+                .append(value._1())
+                .append(":")
+                .append(value._2())
+                .append(",");
+          }
+          return new Tuple2<>(pair._1(), concatenated.toString());
+        })
+        .repartition(20);
+    JavaRDD<String> strings = mappedPairs
+        .mapToPair(tuple -> new Tuple2<>(tuple._2(), tuple._1()))
+        .reduceByKey(Double::sum)
+        .filter(tuple -> tuple._2() > 0.5)
+        .sortByKey()
+        .map(t -> t._1);
+
+    wordCountPairs1 = strings.mapToPair(word -> new Tuple2<>(word, 1))
+        .reduceByKey(Integer::sum)
+        .mapToPair(tuple ->
+            new Tuple2<>(tuple._1(), new Tuple2<>(tuple._1().length(), tuple._2())));
+    joinedPairs = wordCountPairs1
+        .cartesian(wordCountPairs2)
+        .filter(tuple -> tuple._1()._1().compareTo(tuple._2()._1()) < 0)
+        .mapToPair(tuple -> new Tuple2<>(
+            new Tuple2<>(tuple._1()._1(), tuple._2()._1()),
+            new Tuple2<>(
+                tuple._1()._2()._1() + tuple._2()._2()._1(),
+                tuple._1()._2()._2() + tuple._2()._2()._2())));
+    sortedPairs = joinedPairs
+        .filter(tuple -> tuple._2()._2() > 3)
+        .mapValues(tuple -> (double) tuple._1() / tuple._2())
+        .mapToPair(pair -> new Tuple2<>(pair._2(), pair._1()))
+        .repartition(10)
+        .sortByKey(false);
+    sortedPairs.mapPartitionsWithIndex(
+        (index, iter) -> {
+          List<Tuple2<Double, Tuple2<String, String>>> tuples = new ArrayList<>();
+          while (iter.hasNext()) {
+            tuples.add(new Tuple2<>(
+                iter.next()._1(),
+                new Tuple2<>(
+                    index + "-" + iter.next()._2()._1(),
+                    iter.next()._2()._2())));
+          }
+          return tuples.iterator();
+        },
+        true);
   }
 
   /**
@@ -44,15 +140,18 @@ public class EndToEnd {
    * @since 1.0.0
    */
   public static void main(String[] args) {
-    SparkConf conf = new SparkConf().setAppName("SystemTest").setMaster("local[1]");
-    conf.set("spark.plugins", "com.asml.apa.wta.spark.WtaPlugin");
+    SparkConf conf = new SparkConf()
+        .setAppName("SystemTest")
+        .setMaster("local")
+        .set("spark.sql.shuffle.partitions", "500")
+        .set("spark.plugins", "com.asml.apa.wta.spark.WtaPlugin");
     System.setProperty("configFile", args[0]);
+    //    System.setProperty("configFile", "adapter/spark/src/test/resources/config.json");
     SparkSession spark = SparkSession.builder().config(conf).getOrCreate();
     SparkContext sc = spark.sparkContext();
-    testFile = JavaSparkContext.fromSparkContext(sc).textFile(args[1]);
-    for (int i = 0; i < 10; i++) {
-      invokeJob();
-    }
+    sparkOperation(JavaSparkContext.fromSparkContext(sc).textFile(args[1]));
+    //    sparkOperation(
+    //        JavaSparkContext.fromSparkContext(sc).textFile("adapter/spark/src/test/resources/e2e-input.txt"));
     sc.stop();
   }
 }